Refactoring: Separated UploadedImages
