+++ client/src/admin/components/ProductPage/ProductEdit/UploadedImages.jsx/index.module.scss
@@ -1,0 +1,6 @@
+.images {
+  img {
+    height: 100px;
+    object-fit: cover;
+  }
+}
