+++ client/src/admin/components/ProductPage/ProductEdit/index.module.scss
@@ -1,0 +1,9 @@
+.images {
+  display: flex;
+  gap: 25px;
+
+  img {
+    height: 100px;
+    object-fit: cover;
+  }
+}
